Fix escaping of SMTP_AUTH_MECHANISM in env template
Correctly escapes the quotation marks of the SMTP_AUTH_MECHANISM
variable in .env.template. The file is sourced by the shell and
swallows the simple quotation marks. However, we need a proper JSON
value, that is, it has to have quotation marks. This is fixed by adding
escaped quotation marks.