release optimization
This optimization is not only for size but also runtime. Who doesn't want to use less CPU cycles? ;-)

Since releases are not done that often and usually via build machines and pipelines, I don't see it as an issue that the build time doubled.

| Description         | Change | Before   | After     |
|---------------------|--------|----------|-----------|
| Build time          | +110%  | 5min 31s | 11min 34s |
| Binary size (bytes) |  -42%  | 49466432 | 28576480  |

Build command: `cargo build --features sqlite,mysql,postgresql --release`